agents/langgraph/requirements.txt
langgraph>=0.2.0
openai>=1.40.0
pydantic>=2.7
python-dotenv>=1.0